fixed some problem with redirect
