+++ project/appweb/src/main/webapp/admin/admin.html
@@ -6,7 +6,7 @@
 </head>
 <body align=center>
 <h1 center> Admin work page </h1>
-    <a href="/myweb/admin/info.html">Info</a>
+    <a href="/myweb/admin/info">Info</a>
 <p center>
     <c></c>
 </p>
