+++ project/appweb/src/main/webapp/admin/info.html
@@ -7,7 +7,7 @@
 <body align=center>
 <h1 center> Admin work page </h1>
 <h1> Please enter kind of person </h1>
-<form action="info" method="post">
+<form action="/myweb/admin/info" method="post">
     <input name="kind" placeholder=kind type=text> <br/>
     <input type=submit value=OK>
 </form>
